+++ packages/jimp/package.json
@@ -67,6 +67,7 @@
     "@jimp/plugin-resize": "0.3.5",
     "@jimp/plugin-blit": "0.3.5",
     "@jimp/plugin-rotate": "0.3.5",
+    "@jimp/plugin-print": "0.3.5",
     "@jimp/plugin-crop": "0.3.5",
     "@jimp/plugin-autocrop": "0.3.5"
   },
