+++ packages/plugin-print/README.md
@@ -1,0 +1,3 @@
+# @jimp/plugin-print
+
+print an image.
