+++ packages/plugin-print/fonts/open-sans/open-sans-10-black/open-sans-10-black.png
