+++ packages/plugin-print/fonts/open-sans/open-sans-12-black/open-sans-12-black.fnt
