+++ packages/plugin-print/fonts/open-sans/open-sans-14-black/open-sans-14-black.png
