+++ packages/plugin-print/fonts/open-sans/open-sans-16-black/open-sans-16-black.fnt
