+++ packages/plugin-print/fonts/open-sans/open-sans-16-white/open-sans-16-white.png
