+++ packages/plugin-print/fonts/open-sans/open-sans-32-white/open-sans-32-white.fnt
